+++ ididi/__init__.py
@@ -14,7 +14,7 @@ copyright: (c) 2024 by race chan.
 license: MIT, see LICENSE for more details.
 """
 
-VERSION = "1.3.4"
+VERSION = "1.3.5"
 
 __version__ = VERSION